Revert "Revert "重构代码和优化工作流：引入自动格式化、拆分组件和修正命名""
