+++ src/All.js
@@ -1,7 +1,9 @@
 import { Message } from "@arco-design/web-react";
-import Content from "./components/Content";
+
 import { getCurrentUser } from "./apis";
 import { thunder } from "./apis/axios";
+import Content from "./components/Content";
+import { ContentProvider } from "./components/ContentContext";
 
 export default function All() {
   const getEntries = async (offset = 0, status = null) => {
@@ -34,10 +36,12 @@ export default function All() {
   };
 
   return (
-    <Content
-      info={{ from: "all", id: "" }}
-      getEntries={getEntries}
-      markAllAsRead={markAllAsRead}
-    />
+    <ContentProvider>
+      <Content
+        info={{ from: "all", id: "" }}
+        getEntries={getEntries}
+        markAllAsRead={markAllAsRead}
+      />
+    </ContentProvider>
   );
 }
