+++ src/apis/index.js
@@ -1,5 +1,6 @@
-import { thunder } from "./axios";
 import { Message } from "@arco-design/web-react";
+
+import { thunder } from "./axios";
 
 export async function updateEntryStatus(entry, status = "toggle") {
   let newStatus;
