+++ src/components/AddFeedModal.js
@@ -6,16 +6,17 @@ import {
   Select,
   Switch,
 } from "@arco-design/web-react";
-import { addFeed } from "../apis";
 import { useState } from "react";
+
 import { useStore } from "../Store";
+import { addFeed } from "../apis";
 
 export default function AddFeedModal() {
   const { visible, setVisible, groups, initData } = useStore();
   const [feedModalLoading, setFeedModalLoading] = useState(false);
   const [feedForm] = Form.useForm();
 
-  const handelAddFeed = async (feed_url, group_id, is_full_text) => {
+  const handleAddFeed = async (feed_url, group_id, is_full_text) => {
     setFeedModalLoading(true);
     const response = await addFeed(feed_url, group_id, is_full_text);
     if (response) {
@@ -44,7 +45,7 @@ export default function AddFeedModal() {
         form={feedForm}
         layout="vertical"
         onSubmit={(values) =>
-          handelAddFeed(values.url, values.group, values.crawler)
+          handleAddFeed(values.url, values.group, values.crawler)
         }
         labelCol={{ span: 7 }}
         wrapperCol={{ span: 17 }}
