+++ src/components/Header.js
@@ -18,6 +18,7 @@ import {
   IconUser,
 } from "@arco-design/web-react/icon";
 import { useNavigate } from "react-router-dom";
+
 import { useStore } from "../Store";
 import { applyColor } from "../utils/Colors";
 
@@ -25,7 +26,7 @@ export default function Header({ theme }) {
   const navigate = useNavigate();
   const { toggleTheme, setVisible } = useStore();
 
-  const handelLogout = () => {
+  const handleLogout = () => {
     localStorage.removeItem("server");
     localStorage.removeItem("token");
     localStorage.removeItem("theme");
@@ -97,7 +98,7 @@ export default function Header({ theme }) {
                   />
                   Settings
                 </Menu.Item>
-                <Menu.Item key="1" onClick={handelLogout}>
+                <Menu.Item key="1" onClick={handleLogout}>
                   <IconPoweroff
                     style={{
                       marginRight: 8,
