+++ src/components/ImageWithLazyLoading.css
@@ -1,3 +1,3 @@
 .read {
-    filter: brightness(60%);
+  filter: brightness(60%);
 }
